# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "chess-puzzle-api",
    "description": "Random chess puzzles by rating and theme, from the Lichess puzzle database (CC0). The `fen` of a puzzle is the position *before* the opponent's move: apply `initialMove` to reach the position to solve. Solutions are served only by the solution endpoint.",
    "contact": {
      "name": "Mauri Ulloa",
      "url": "https://mauriulloa.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/puzzles/random": {
      "get": {
        "tags": [
          "puzzles"
        ],
        "operationId": "random",
        "parameters": [
          {
            "name": "rating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "tolerance",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "ratingMin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "ratingMax",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "themes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "themesMode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "excludeThemes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "opening",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxPieces",
            "in": "query",
            "description": "Only positions with at most this many pieces on the board, kings\nincluded.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "board",
            "in": "query",
            "description": "Include the position drawn as text. Off by default because it is by\nfar the largest field in the response.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A random puzzle. With `count`, a batch envelope instead.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PuzzleResponse"
                }
              }
            }
          },
          "400": {
            "description": "A parameter was invalid or unrecognised",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "The API key is unknown or revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "The filters describe an empty slice of the dataset",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "The search ran past the query timeout and was stopped",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/puzzles/{id}": {
      "get": {
        "tags": [
          "puzzles"
        ],
        "operationId": "by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Lichess puzzle id, e.g. `00008`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "board",
            "in": "query",
            "description": "Include the position drawn as text.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The puzzle, without its solution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PuzzleResponse"
                }
              }
            }
          },
          "400": {
            "description": "A parameter was invalid or unrecognised",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "The API key is unknown or revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No puzzle carries that id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/puzzles/{id}/solution": {
      "get": {
        "tags": [
          "puzzles"
        ],
        "operationId": "solution",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Lichess puzzle id, e.g. `00008`",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The moves that solve the puzzle",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolutionResponse"
                }
              }
            }
          },
          "400": {
            "description": "A parameter was invalid or unrecognised",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "The API key is unknown or revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No puzzle carries that id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stats": {
      "get": {
        "tags": [
          "reference"
        ],
        "operationId": "stats",
        "responses": {
          "200": {
            "description": "Dataset size, rating distribution and provenance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          },
          "401": {
            "description": "The API key is unknown or revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/themes": {
      "get": {
        "tags": [
          "reference"
        ],
        "operationId": "themes",
        "responses": {
          "200": {
            "description": "Every theme, with how many puzzles carry it and what it means",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThemesResponse"
                }
              }
            }
          },
          "401": {
            "description": "The API key is unknown or revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/usage": {
      "get": {
        "tags": [
          "reference"
        ],
        "operationId": "usage",
        "responses": {
          "200": {
            "description": "Aggregate usage over the last 30 days",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                }
              }
            }
          },
          "401": {
            "description": "The API key is unknown or revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "The shape of every error response.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Machine-readable error kind, e.g. `bad_request` or `rate_limited`."
          },
          "hint": {
            "type": [
              "string",
              "null"
            ],
            "description": "How to fix it, when there is something useful to say."
          },
          "message": {
            "type": "string",
            "description": "What went wrong, in plain language."
          }
        }
      },
...
      "PuzzleResponse": {
        "type": "object",
        "required": [
          "id",
          "fen",
          "initialMove",
          "solverColor",
          "rating",
          "ratingDeviation",
          "popularity",
          "nbPlays",
          "pieces",
          "themes",
          "openingTags"
        ],
        "properties": {
          "analysisUrl": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opens the position on a real board, for a person."
          },
          "board": {
            "type": [
              "string",
              "null"
            ],
            "description": "The position drawn as text. Present only when `board=true` is asked\nfor, since it is by far the largest field."
          },
          "fen": {
            "type": "string",
            "description": "The position as Lichess stores it: *before* the opponent's move."
          },
          "gameUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "initialMove": {
            "type": "string",
            "description": "The opponent's move, in UCI. Apply it to `fen` to reach the position\nthe player is asked to solve."
          },
          "initialMoveSan": {
            "type": [
              "string",
              "null"
            ],
            "description": "The opponent's move in SAN, e.g. `Bxg3`. Chess literature is written\nin SAN, so this is what people and language models read fluently."
          },
          "nbPlays": {
            "type": "integer",
            "format": "int64"
          },
          "openingTags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pieces": {
            "type": "integer",
            "format": "int32",
            "description": "Pieces on the board the player solves, kings included.",
            "minimum": 0
          },
          "popularity": {
            "type": "integer",
            "format": "int64"
          },
          "positionFen": {
            "type": [
              "string",
              "null"
            ],
            "description": "The position *after* `initialMove`: what the player actually faces.\n`fen` is kept as the dataset stores it so existing clients do not break."
          },
          "rating": {
            "type": "integer",
            "format": "int64"
          },
          "ratingDeviation": {
            "type": "integer",
            "format": "int64"
          },
          "solverColor": {
            "type": "string",
            "description": "The colour the solver plays, already accounting for `initialMove`."
          },
          "themes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
...
}